Move authz models to rocketchat:models
