--- packages/rocketchat-authorization/server/models/Users.js
@@ -1,15 +1,0 @@
-import { RocketChat } from 'meteor/rocketchat:lib';
-
-RocketChat.models.Users.roleBaseQuery = function(userId) {
-	return { _id: userId };
-};
-
-RocketChat.models.Users.findUsersInRoles = function(roles, scope, options) {
-	roles = [].concat(roles);
-
-	const query = {
-		roles: { $in: roles },
-	};
-
-	return this.find(query, options);
-};
