+++ packages/rocketchat-lib/package.js
@@ -155,6 +155,8 @@ Package.onUse(function(api) {
 	api.addFiles('server/models/Users.js', 'server');
 	api.addFiles('server/models/ExportOperations.js', 'server');
 	api.addFiles('server/models/UserDataFiles.js', 'server');
+	api.addFiles('server/models/Permissions.js', 'server');
+	api.addFiles('server/models/Roles.js', 'server');
 
 	api.addFiles('server/oauth/oauth.js', 'server');
 	api.addFiles('server/oauth/facebook.js', 'server');
