+++ packages/rocketchat-lib/server/models/Permissions.js
@@ -1,0 +1,3 @@
+import { Permissions } from 'meteor/rocketchat:models';
+
+RocketChat.models.Permissions = Permissions;
