+++ packages/rocketchat-lib/server/models/Roles.js
@@ -1,0 +1,3 @@
+import { Roles } from 'meteor/rocketchat:models';
+
+RocketChat.models.Roles = Roles;
