+++ packages/rocketchat-models/server/models/Subscriptions.js
@@ -1,7 +1,8 @@
 import { Base } from './_Base';
 import { Match } from 'meteor/check';
 import Rooms from './Rooms';
-import { getDefaultSubscriptionPref } from 'meteor/rocketchat:utils';
+import Users from './Users';
+import _ from 'underscore';
 
 export class Subscriptions extends Base {
 	constructor(...args) {
@@ -28,6 +29,39 @@ export class Subscriptions extends Base {
 		this.tryEnsureIndex({ 'userHighlights.0': 1 }, { sparse: 1 });
 	}
 
+	roleBaseQuery(userId, scope) {
+		if (scope == null) {
+			return;
+		}
+
+		const query = { 'u._id': userId };
+		if (!_.isUndefined(scope)) {
+			query.rid = scope;
+		}
+		return query;
+	}
+
+	findUsersInRoles(roles, scope, options) {
+		roles = [].concat(roles);
+
+		const query = {
+			roles: { $in: roles },
+		};
+
+		if (scope) {
+			query.rid = scope;
+		}
+
+		const subscriptions = this.find(query).fetch();
+
+		const users = _.compact(_.map(subscriptions, function(subscription) {
+			if ('undefined' !== typeof subscription.u && 'undefined' !== typeof subscription.u._id) {
+				return subscription.u._id;
+			}
+		}));
+
+		return Users.find({ _id: { $in: users } }, options);
+	}
 
 	// FIND ONE
 	findOneByRoomIdAndUserId(roomId, userId, options) {
